+++ R/deprec-do.r
@@ -1,9 +1,9 @@
 #' Do anything
 #'
 #' @description
-#' \Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
+#' \Sexpr[results=rd, stage=render]{lifecycle::badge("retired")}
 #'
-#' `do()` is deprecated as of dplyr 1.0.0, because its syntax never really
+#' `do()` is superseded as of dplyr 1.0.0, because its syntax never really
 #' felt like it belong with the rest of dplyr. It's replaced by a combination
 #' of [summarise()] (which can now produce multiple rows and multiple columns),
 #' [condense()] (which creates a [rowwise] tibble containing list-columns),
@@ -46,10 +46,6 @@
 #'   models %>% summarise(broom::tidy(mod))
 #' }
 do <- function(.data, ...) {
-  lifecycle::deprecate_warn("1.0.0", "do()",
-    details = "Use condense() or summarise()"
-  )
-
   UseMethod("do")
 }